+++ src/algorithm/mod.rs
@@ -5,8 +5,8 @@ pub mod convexhulltrick;
 pub mod cumsum;
 pub mod enumdivisors;
 pub mod eratosthenes;
+pub mod fasteratosthenes;
 pub mod inversionnumber;
 pub mod largestrectangle;
-pub mod linearsieve;
 pub mod prime;
 pub mod run_length;
